Moved the job logs into the job container.
