+++ client/job-view/jobView.pug
@@ -1,3 +1,5 @@
 div#job-view
 
   div(data-hook="job-results-container")
+
+  div(data-hook="job-info-container")
