+++ client/job-view/templates/jobLogsView.pug
