+++ client/templates/pages/workflowManager.pug
@@ -95,8 +95,6 @@ section.page
 
   div(data-hook="active-job-container")
 
-  div(data-hook="workflow-info-container")
-
   div#review-settings-section.card(style="display: none;")
 
     div.card-header
